Text transformer updated
