+++ StoryLine.Utils/Actions/TextTransformAction.cs
@@ -5,16 +5,16 @@ using StoryLine.Utils.Contracts;
 
 namespace StoryLine.Utils.Actions
 {
-    public class TransformationAction : IAction
+    public class TextTransformAction : IAction
     {
         private readonly Type _type;
-        private readonly ITextProvider _provider;
+        private readonly string _text;
         private readonly ITextConverter _converter;
 
-        public TransformationAction(Type type, ITextProvider provider, ITextConverter converter)
+        public TextTransformAction(Type type, string text, ITextConverter converter)
         {
             _type = type ?? throw new ArgumentNullException(nameof(type));
-            _provider = provider ?? throw new ArgumentNullException(nameof(provider));
+            _text = text ?? throw new ArgumentNullException(nameof(text));
             _converter = converter ?? throw new ArgumentNullException(nameof(converter));
         }
 
@@ -23,13 +23,9 @@ namespace StoryLine.Utils.Actions
             if (actor == null)
                 throw new ArgumentNullException(nameof(actor));
 
-            var text = _provider.GetText();
-            if (string.IsNullOrEmpty(text))
-                throw new ExpectationException($"Failed to get text for artifact of type {_type}.");
-
-            var result = _converter.Convert(text, _type);
+            var result = _converter.Convert(_text, _type);
             if (result == null)
-                throw new ExpectationException($"Failed to convert text into artifact of type {_type}. Text to convert: {text}.");
+                throw new ExpectationException($"Failed to convert text into artifact of type {_type}. Text to convert: {_text}.");
 
             actor.Artifacts.Add(result);
         }
