--- StoryLine.Utils/Contracts/ITextProvider.cs
@@ -1,7 +1,0 @@
-﻿namespace StoryLine.Utils.Contracts
-{
-    public interface ITextProvider
-    {
-        string GetText();
-    }
-}
